requirements.txt
GPUtil==1.4.0
Pillow==9.1.0
psutil==5.9.0
py-cpuinfo==8.0.0
PyYAML==6.0
ttkthemes==3.2.2
